// src/app-react/components/dashboard-visualization/components/Histogram/types.ts
export type ChartData = {
  indexes: Float64Array;
  indexSelectionDetails: IndexSelectionDetails;
  selectedIndex: number | null;
  xAxisLabel: string;
  yAxisLabel: string;
  range: Range | null;
  fullRange: Range;
  dataSetId: string | null;
  refs: ChartRefs | null;
  data: BinData | null;
  allData: HistogramData[];
  allDataMaxY: number;
  allDataMinY: number;
  datasetLabels: string[];
  channelNames: string[];
  selectedChannel: string | null;
};
// ...
export type ChartRefs = {
  chartSurface: SciChartSurface;
  columnSeries: FastColumnRenderableSeries;
  lineSeries: FastLineRenderableSeries;
  pointSeries: FastLineRenderableSeries;
  overviewColumnSeries: FastColumnRenderableSeries;
  overviewLineSeries: FastLineRenderableSeries;
  overviewPointSeries: FastLineRenderableSeries;
  xAxis: NumericAxis;
  yAxis: NumericAxis;
  wasmContext: TSciChart;
  overview: SciChartOverview;
};

export type Bin = {
  start: number;
  midPoint: number;
  end: number;
  y: number;
};

type BinData = {
  bins: Bin[];
  xDomain: Range;
  binWidth: number;
  y: number[];
};
// ...
export type HistogramData = {
  chanName: string;
  type: "h2axis";
  data: {
    x: {
      start: number;
      midPoint: number;
      end: number;
      label: string;
    }[];
    xUnits: string;
    xLabel: string;
    y: number[];
    yUnits: string;
    yLabel: string;
  };
};